#35 Querying.Storage.Tests recovered
