+++ Querying/Querying.Storage/Querying.Storage.csproj
@@ -84,6 +84,7 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DbModelBuilderConfigurator.cs" />
     <Compile Include="DbModelBuilderExtensions.cs" />
     <Compile Include="DbModelFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
