+++ Querying/Tests/Querying.Storage.Tests/packages.config
@@ -4,6 +4,7 @@
   <package id="2GIS.NuClear.Utils" version="1.0.3" targetFramework="net461" />
   <package id="Effort.EF6" version="1.1.4" targetFramework="net461" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net461" />
+  <package id="Moq" version="4.2.1510.2205" targetFramework="net461" />
   <package id="NMemory" version="1.0.1" targetFramework="net461" />
-  <package id="NUnit" version="3.2.0" targetFramework="net461" />
+  <package id="NUnit" version="2.6.4" targetFramework="net461" />
 </packages>
